broke up the original code into two files
