+++ client/components/WineList/winesList.js
@@ -2,7 +2,7 @@ import React from 'react'
 import {connect} from 'react-redux'
 import {getWines} from '../../store/allWines'
 import WinesIcon from './winesIcon'
-import SortFilter from '../SortFilter'
+import SortFilter from '../SortFilter/Grid'
 
 export class WinesList extends React.Component {
   constructor() {
